requirements.txt
python-binance
coinbase
python-telegram-bot
selenium
certifi